Introduce MediaViewerFlickToDismiss and extract to its own file
